# sdks/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "LiteGen API",
        "description": "Universal proxy for AI image and video generation.",
        "license": {
            "name": "MIT",
            "identifier": "MIT"
        },
        "version": "0.1.0"
    },
    "paths": {
...
        "/v1/images/generations": {
            "post": {
                "tags": [
                    "Images"
                ],
                "summary": "POST /v1/images/generations \u2014 Generate images (OpenAI-compatible).",
                "operationId": "generate_image",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ImageGenerationRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Image generated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ImageGenerationResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "502": {
                        "description": "Provider error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "BaseGenerationRequest": {
                "type": "object",
                "required": [
                    "prompt",
                    "model"
                ],
                "properties": {
                    "extra": {},
                    "metadata": {},
                    "model": {
                        "type": "string"
                    },
                    "n": {
                        "type": "integer",
                        "format": "int32",
                        "minimum": 0
                    },
                    "negative_prompt": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "prompt": {
                        "type": "string"
                    },
                    "reference_images": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ReferenceImage"
                        }
                    },
                    "seed": {
                        "type": [
                            "integer",
                            "null"
                        ],
                        "format": "int64"
                    },
                    "strict": {
                        "type": "boolean"
                    }
                }
            },
...
            "CostSource": {
                "type": "string",
                "enum": [
                    "dynamic",
                    "estimated"
                ]
            },
...
            "ErrorDetail": {
                "type": "object",
                "required": [
                    "message",
                    "type"
                ],
                "properties": {
                    "code": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "message": {
                        "type": "string"
                    },
                    "provider_error": {},
                    "type": {
                        "type": "string"
                    }
                }
            },
            "ErrorResponse": {
                "type": "object",
                "required": [
                    "error"
                ],
                "properties": {
                    "error": {
                        "$ref": "#/components/schemas/ErrorDetail"
                    }
                }
            },
...
            "ImageGenerationRequest": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/BaseGenerationRequest"
                    },
                    {
                        "type": "object",
                        "properties": {
                            "aspect_ratio": {
                                "type": [
                                    "string",
                                    "null"
                                ]
                            },
                            "guidance_scale": {
                                "type": [
                                    "number",
                                    "null"
                                ],
                                "format": "double"
                            },
                            "quality": {
                                "type": [
                                    "string",
                                    "null"
                                ]
                            },
                            "response_format": {
                                "type": "string"
                            },
                            "size": {
                                "type": [
                                    "string",
                                    "null"
                                ]
                            },
                            "steps": {
                                "type": [
                                    "integer",
                                    "null"
                                ],
                                "format": "int32",
                                "minimum": 0
                            },
                            "strength": {
                                "type": [
                                    "number",
                                    "null"
                                ],
                                "format": "double"
                            },
                            "style": {
                                "type": [
                                    "string",
                                    "null"
                                ]
                            }
                        }
                    }
                ]
            },
            "ImageGenerationResponse": {
                "type": "object",
                "description": "Response for an image generation request.\nFollows OpenAI's images response format for compatibility.",
                "required": [
                    "created",
                    "data",
                    "model",
                    "provider",
                    "id"
                ],
                "properties": {
                    "created": {
                        "type": "integer",
                        "format": "int64",
                        "description": "Unix timestamp of when the request was created."
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ImageResult"
                        },
                        "description": "Array of generated images."
                    },
                    "id": {
                        "type": "string",
                        "description": "Unique request ID for tracking."
                    },
                    "model": {
                        "type": "string",
                        "description": "The model that was used."
                    },
                    "provider": {
                        "type": "string",
                        "description": "The provider that handled the request."
                    },
                    "usage": {
                        "oneOf": [
                            {
                                "type": "null"
                            },
                            {
                                "$ref": "#/components/schemas/UsageInfo",
                                "description": "Cost information."
                            }
                        ]
                    }
                }
            },
            "ImageResult": {
                "type": "object",
                "description": "A single generated image in the response.",
                "required": [
                    "index"
                ],
                "properties": {
                    "b64_json": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "description": "Base64-encoded image data (if response_format=b64_json)."
                    },
                    "content_type": {
                        "type": "string",
                        "description": "Content type of the image (e.g. \"image/png\")."
                    },
                    "index": {
                        "type": "integer",
                        "format": "int32",
                        "description": "Index in the batch.",
                        "minimum": 0
                    },
                    "revised_prompt": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "description": "Revised prompt (if provider modified the prompt, e.g. DALL-E 3)."
                    },
                    "url": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "description": "URL of the generated image (if response_format=url)."
                    }
                }
            },
...
            "RefImageKind": {
                "type": "string",
                "enum": [
                    "base64",
                    "url",
                    "blob"
                ]
            },
...
            "ReferenceImage": {
                "type": "object",
                "required": [
                    "type",
                    "value"
                ],
                "properties": {
                    "role": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "type": {
                        "$ref": "#/components/schemas/RefImageKind"
                    },
                    "value": {
                        "type": "string"
                    }
                }
            },
...
            "UsageInfo": {
                "type": "object",
                "description": "Cost / usage information for a generation.",
                "required": [
                    "cost_usd",
                    "tokens",
                    "cost_source"
                ],
                "properties": {
                    "cost_source": {
                        "$ref": "#/components/schemas/CostSource",
                        "description": "Cost source: \"dynamic\" (from API) or \"estimated\" (from pricing table)."
                    },
                    "cost_usd": {
                        "type": "number",
                        "format": "double",
                        "description": "Provider cost in USD."
                    },
                    "tokens": {
                        "type": "integer",
                        "format": "int64",
                        "description": "Internal token cost.",
                        "minimum": 0
                    }
                }
            },
...
        }
    },
...
}